+++ C++/Navio/Common/RCOutput.h
@@ -1,12 +1,12 @@
 #pragma once
 
-#include <cinttypes>
+#include <cstddef>
 
 class RCOutput
 {
 public:
-  virtual bool initialize(const uint32_t& channel) = 0;
-  virtual bool enable(const uint32_t& channel) = 0;
-  virtual bool setFrequency(const uint32_t& channel, const uint32_t& frequency) = 0;
-  virtual bool setDutyCycle(const uint32_t& channel, const double& period_us) = 0;
+  virtual bool initialize(const size_t& channel) = 0;
+  virtual bool enable(const size_t& channel) = 0;
+  virtual bool setFrequency(const size_t& channel, const size_t& frequency) = 0;
+  virtual bool setDutyCycle(const size_t& channel, const double& period_us) = 0;
 };
