+++ C++/Navio/Navio+/RCOutput_Navio.h
@@ -7,10 +7,10 @@ class RCOutput_Navio : public RCOutput
 {
 public:
   explicit RCOutput_Navio();
-  bool initialize(const uint32_t&) override;
-  bool enable(const uint32_t&) override;
-  bool setFrequency(const uint32_t&, const uint32_t& frequency) override;
-  bool setDutyCycle(const uint32_t& channel, const double& period) override;
+  bool initialize(const size_t&) override;
+  bool enable(const size_t&) override;
+  bool setFrequency(const size_t&, const size_t& frequency) override;
+  bool setDutyCycle(const size_t& channel, const double& period) override;
 
 private:
   PCA9685 pwm_;
